+++ swap/src/tests/mod.rs
@@ -3,4 +3,4 @@ mod happy_path_restart_alice;
 mod happy_path_restart_bob_after_comm;
 mod punish;
 mod refund_restart_alice;
-pub mod testutils;
+pub mod utils;
